+++ cmd/server/main.go
@@ -2,155 +2,39 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"os"
-	"sync"
-	"time"
 
-	"resty.dev/v3"
-
-	"github.com/go-co-op/gocron/v2"
-	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/joho/godotenv"
-
-	"github.com/cloudsftp/botificator/pkg/analyzer"
-	"github.com/cloudsftp/botificator/pkg/db"
-	"github.com/cloudsftp/botificator/pkg/load"
-	"github.com/cloudsftp/botificator/pkg/notificator"
+	"github.com/cloudsftp/botificator/pkg/server"
+	"github.com/sirupsen/logrus"
 )
-
-var startTime = time.Date(2024, 6, 1, 0, 0, 0, 0, time.UTC)
 
 func main() {
 	ctx := context.Background()
 
-	server, err := NewServer(ctx)
+	logLevelStr := os.Getenv("LOG_LEVEL")
+	if logLevelStr == "" {
+		logLevelStr = "trace"
+	}
+
+	logLevel, err := logrus.ParseLevel(logLevelStr)
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "error in setup: %s\n", err)
-		os.Exit(1)
+		logrus.Fatalf("Invalid log level: %s", logLevelStr)
+	}
+
+	logrus.SetLevel(logLevel)
+	logrus.SetFormatter(&logrus.TextFormatter{
+		FullTimestamp:   true,
+		TimestampFormat: "2006-01-02 15:04:05",
+	})
+
+	server, err := server.New(ctx)
+	if err != nil {
+		logrus.Fatalf("error in setup: %s", err)
 	}
 	defer server.Close()
 
 	err = server.Run(ctx)
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "error while running: %s\n", err)
-		os.Exit(1)
-	}
-}
-
-type Server struct {
-	notificator  *notificator.Notificator
-	pool         *pgxpool.Pool
-	client       *resty.Client
-	scheduler    gocron.Scheduler
-	errors       chan error
-	databaseLock *sync.RWMutex
-}
-
-func NewServer(ctx context.Context) (*Server, error) {
-	err := godotenv.Load()
-	if err != nil {
-		return nil, fmt.Errorf("could not load environment: %w\n", err)
-	}
-
-	notificator, err := notificator.New()
-	if err != nil {
-		return nil, fmt.Errorf("could not create notificator: %w\n", err)
-	}
-
-	pool, err := db.SetupDatabase(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not setup database: %w\n", err)
-	}
-
-	client := resty.New()
-
-	scheduler, err := gocron.NewScheduler()
-	if err != nil {
-		return nil, fmt.Errorf("could not create scheduler: %w\n", err)
-	}
-
-	err = notificator.SendMessageDeployed(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not send message: %w\n", err)
-	}
-
-	errors := make(chan error)
-	var databaseLock sync.RWMutex
-
-	return &Server{notificator, pool, client, scheduler, errors, &databaseLock}, nil
-}
-
-func (s *Server) Close() {
-	s.pool.Close()
-	s.client.Close()
-}
-
-func (s *Server) Run(ctx context.Context) error {
-	s.UpdateDatabase(ctx)
-	_, err := s.scheduler.NewJob(
-		gocron.DurationJob(15*time.Minute),
-		gocron.NewTask(s.UpdateDatabase, ctx),
-	)
-	if err != nil {
-		return fmt.Errorf("could not set up database cron job: %w", err)
-	}
-
-	_, err = s.scheduler.NewJob(
-		gocron.DailyJob(1, gocron.NewAtTimes(gocron.NewAtTime(5, 0, 0))),
-		gocron.NewTask(s.SendUpdate, ctx),
-	)
-	if err != nil {
-		return fmt.Errorf("could not set up notification cron job: %w", err)
-	}
-
-	s.scheduler.Start()
-	defer func() {
-		err := s.scheduler.Shutdown()
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "could not shutdown scheduler: %s", err)
-		}
-	}()
-
-	for {
-		select {
-		case err := <-s.errors:
-			fmt.Fprintf(os.Stderr, "runtime error: %s", err)
-		}
-	}
-}
-
-func (s *Server) UpdateDatabase(ctx context.Context) {
-	ok := s.databaseLock.TryLock()
-	if !ok {
-		// TODO: warn being updated already
-		return
-	}
-	defer s.databaseLock.Unlock()
-
-	// s.notificator.SendMessage(ctx, "Updating Database...")
-
-	err := load.LoadDataIntoDatabase(ctx, s.client, s.pool, startTime)
-	if err != nil {
-		s.errors <- fmt.Errorf("could not load data into database: %s\n", err)
-	}
-}
-
-func (s *Server) SendUpdate(ctx context.Context) {
-	ok := s.databaseLock.TryRLock()
-	if !ok {
-		// TODO: wait for an hour, otherwise error out
-		return
-	}
-	defer s.databaseLock.RUnlock()
-
-	averages, err := db.GetMovingAverages(ctx, s.pool)
-	if err != nil {
-		s.errors <- fmt.Errorf("could not get moving averages: %s\n", err)
-	}
-
-	err = analyzer.Analyze(averages)
-	if err != nil {
-		s.errors <- fmt.Errorf("could not analyze averages: %s\n", err)
+		logrus.Fatalf("error while running: %s", err)
 	}
 }
