+++ pkg/db/db.go
@@ -3,12 +3,13 @@ package db
 import (
 	"context"
 	"fmt"
-	"os"
 	"strconv"
 	"time"
 
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/sirupsen/logrus"
+	log "github.com/sirupsen/logrus"
 
 	"github.com/cloudsftp/botificator/pkg/analyzer"
 	"github.com/cloudsftp/botificator/pkg/api"
@@ -18,96 +19,6 @@ const (
 	ohclTable        = "btc_5min"
 	dailyAverageView = "btc_daily_avg"
 )
-
-func SetupDatabase(ctx context.Context) (*pgxpool.Pool, error) {
-	pool, err := createConnectionPool(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("could not create a connection pool: %w", err)
-	}
-
-	err = createTables(ctx, pool)
-	if err != nil {
-		return nil, fmt.Errorf("could not create tables: %w", err)
-	}
-
-	return pool, nil
-}
-
-func createConnectionPool(ctx context.Context) (*pgxpool.Pool, error) {
-	pass := os.Getenv("DB_PASS")
-	if pass == "" {
-		return nil, fmt.Errorf("no environment variable DB_PASS")
-	}
-
-	host := os.Getenv("DB_HOST")
-	if host == "" {
-		return nil, fmt.Errorf("no environment variable DB_HOST")
-	}
-
-	port := os.Getenv("DB_PORT")
-	if port == "" {
-		return nil, fmt.Errorf("no environment variable DB_PORT")
-	}
-
-	connectionString := fmt.Sprintf("postgres://postgres:%s@%s:%s/postgres", pass, host, port)
-
-	config, err := pgxpool.ParseConfig(connectionString)
-	if err != nil {
-		return nil, err
-	}
-
-	config.MaxConns = 20
-	config.MinConns = 2
-
-	return pgxpool.NewWithConfig(ctx, config)
-}
-
-func createTables(ctx context.Context, pool *pgxpool.Pool) error {
-	_, err := pool.Exec(ctx, fmt.Sprintf(`
-		CREATE TABLE IF NOT EXISTS %s (
-			time   TIMESTAMPTZ    NOT NULL UNIQUE,
-			open   DECIMAL(30,5)  NOT NULL,
-			high   DECIMAL(30,5)  NOT NULL,
-			low    DECIMAL(30,5)  NOT NULL,
-			close  DECIMAL(30,5)  NOT NULL,
-			volume DECIMAL(40,20) NOT NULL
-		)
-	`, ohclTable))
-	if err != nil {
-		return fmt.Errorf("could not create table %s: %w", ohclTable, err)
-	}
-
-	_, err = pool.Exec(ctx, fmt.Sprintf(`
-		SELECT create_hypertable('%s', 'time', if_not_exists => true)
-	`, ohclTable))
-	if err != nil {
-		return fmt.Errorf("could not create hypertable %s: %w", ohclTable, err)
-	}
-
-	_, err = pool.Exec(ctx, fmt.Sprintf(`
-		CREATE MATERIALIZED VIEW IF NOT EXISTS %s
-		WITH (timescaledb.continuous) AS
-		SELECT
-			time_bucket('1 day', time) AS day,
-			AVG(low) as average
-		FROM
-			%s
-		GROUP BY
-			day
-    `, dailyAverageView, ohclTable))
-	if err != nil {
-		return fmt.Errorf("could not create view %s : %w", dailyAverageView, err)
-	}
-
-	_, err = pool.Exec(ctx, fmt.Sprintf(`
-		ALTER MATERIALIZED VIEW %s set (timescaledb.materialized_only = false);
-    `, dailyAverageView))
-	if err != nil {
-		return fmt.Errorf("could not set %s to real time: %w", dailyAverageView, err)
-	}
-
-	return nil
-}
 
 // GetLatestTimestamp returns the timestamp of the latest row
 func GetLatestTimestamp(ctx context.Context, conn *pgx.Conn) (int64, bool, error) {
@@ -120,7 +31,7 @@ func GetLatestTimestamp(ctx context.Context, conn *pgx.Conn) (int64, bool, error
 
 	result, err := conn.Query(ctx, query)
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "could not execute query to get latest item in %s: %s\n", ohclTable, err)
+		log.Errorf("could not execute query to get latest item in %s: %s", ohclTable, err)
 		return 0, false, err
 	}
 	defer result.Close()
@@ -132,7 +43,7 @@ func GetLatestTimestamp(ctx context.Context, conn *pgx.Conn) (int64, bool, error
 	var latestTimestamp int64
 	err = result.Scan(&latestTimestamp)
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "could not execute query to get values from result: %s\n", err)
+		logrus.Errorf("could not execute query to get values from result: %s", err)
 		return 0, false, err
 	}
 
@@ -160,12 +71,12 @@ func InsertDataPoints(ctx context.Context, conn *pgx.Conn, elements []api.Histor
 	)
 
 	if err != nil {
-		fmt.Fprintf(os.Stderr, "could not execute query to insert rows: %s\n", err)
+		logrus.Errorf("could not execute query to insert rows: %s", err)
 		return false, err
 	}
 
 	if copyCount == 0 {
-		fmt.Fprint(os.Stderr, "no rows inserted")
+		logrus.Error("no rows inserted")
 		return false, nil
 	}
